chore: 자료형 수정

+++ src/main/java/gdg/whowantit/entity/Funding.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public class Funding {
     @Id
-    private int fundingId;
+    private Long fundingId;
 
     private String content;
